💡refactor: problemList request, response 변경에 따른 수정
