+++ src/apis/type.ts
@@ -1,8 +1,15 @@
+// gitLogin.ts 의 타입
 export type GitLoginResponse = {
   message: string;
   statusCode: number;
   token: LoginToken;
   userData: LoginUserData;
+};
+
+export type RefreshTokenResponse = {
+  statusCode: number;
+  message: string;
+  accessToken: string;
 };
 
 export type LoginToken = {
@@ -21,12 +28,7 @@ export type PostRefreshTokenProps = {
   refreshToken: string;
 };
 
-export type RefreshTokenResponse = {
-  statusCode: number;
-  message: string;
-  accessToken: string;
-};
-
+// user.ts 의 타입
 export type UserProfileResponse = {
   statusCode: number;
   message: string;
@@ -37,3 +39,21 @@ export type UserProfileResponse = {
     level: string;
   };
 };
+
+
+// problem.ts 의 타입 
+export type GetProblemListProps = {
+    sector: string;
+    difficulty: number;
+    page: number | boolean;
+    limit: number;
+}
+
+export type GetProblemListResponse = {
+    problemList: {
+        problem_id: number
+        title: string
+        type:string
+    }[]
+    nextPage: number | boolean
+}
